infobase/requirements.txt
python-telegram-bot

langchain
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain-openai
langchain-community
playwright
pypdf
poppler-utils
# python 3.13 not compatible
# tesseract-ocr
langdetect

matplotlib
PyQt6
jupyterlab
tqdm